+++ front/src/components/admin_panel/AdminPanel.jsx
@@ -3,13 +3,15 @@ import React, { Component } from 'react';
 import css from './AdminPanel.module.css';
 
 import { Admin, Resource } from 'react-admin';
-import { PostList, PostEdit, PostCreate, PostIcon } from './posts';
+import { PostList, PostEdit, PostCreate, PostIcon } from './tasks';
 import restProvider from 'ra-data-simple-rest';
+import {ParticipantsCreate, ParticipantsEdit, ParticipantsIcon, ParticipantsList} from "./participants";
 
 function AdminPanel() {
     return (
         <Admin dataProvider={restProvider('http://localhost:3000')}>
-            <Resource name="posts" list={PostList} edit={PostEdit} create={PostCreate} icon={PostIcon} />
+            <Resource name="tasks" list={PostList} edit={PostEdit} create={PostCreate} icon={PostIcon} />
+            <Resource name="participants" list={ParticipantsList} edit={ParticipantsEdit} create={ParticipantsCreate} icon={ParticipantsIcon} />
         </Admin>
     );
 }
